Login and out buttons work
