+++ components/NavigationBar.js
@@ -1,8 +1,24 @@
 import React from 'react'
 import Link from 'next/link'
+import { useRouter } from 'next/router'
+import { deleteMe } from '../utils/auth'
 import { Button, Navbar, Nav } from 'react-bootstrap/'
 
 const NavigationBar = () => {
+  const router = useRouter()
+
+  const handleSignIn = () => {
+    router.push('/api/spotify-auth')
+  }
+
+  const handleSignOut = async () => {
+    try {
+      await deleteMe()
+    } catch (error) {
+      console.log('sign out', error)
+    }
+  }
+
   return (
     <Navbar collapseOnSelect className="navbar-dark bg-dark" expand="lg">
       <Link href="/" passHref>
@@ -23,12 +39,12 @@ const NavigationBar = () => {
         </Nav>
         <Nav className="justify-content-end">
           <Link href="/login" passHref>
-            <Button onClick={() => console.log('sign in')} variant="light" disabled>
+            <Button onClick={() => handleSignIn()} variant="light">
               Login
             </Button>
           </Link>
           <Link href="/login" passHref>
-            <Button onClick={() => console.log('sign out')} variant="light" disabled>
+            <Button onClick={() => handleSignOut()} variant="light">
               Log out
             </Button>
           </Link>
