+++ utils/useAuth.js
@@ -34,26 +34,10 @@ const useAuth = () => {
     }
   }, [router])
 
-  const handleSignIn = () => {
-    router.push('/api/spotify-auth')
-  }
-
-  const handleSignOut = async () => {
-    try {
-      await deleteMe()
-    } catch (error) {
-      console.log('sign out', error)
-    }
-  }
-
   return [
     loaded,
     token,
-    user,
-    {
-      handleSignIn,
-      handleSignOut
-    }
+    user
   ]
 }
